pyproject.toml
[project]
name = "venv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "matplotlib>=3.10.8",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "streamlit>=1.52.2",
]
